refactor: one copy of the session helpers the routes wrote identically
`_save_session_context`, `_record_session_exchange`, `_start_empty_window`
and `_show_handoff` were byte-identical between the openai and anthropic
routes — same bodies, same printed lines — and codex carried a third copy
of `_show_handoff` differing only in a docstring word. `COMPACT_ATTEMPTS`
was declared twice with the same comment. All of it now sits once on
`Base`, under a section saying which routes reach it.

`_session_budget` deliberately stays per-route: the suite monkeypatches
`compact_budget_for` on the chatgptapi and codex modules, so moving the
lookup would move it out from under those tests.
